pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-formatter"
version = "0.1.0"
description = "Agent-driven document transformation system for resumes"
requires-python = ">=3.10"
dependencies = [
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
resume-formatter = "resume_formatter.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["resume_formatter*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
